removing admin api from feature branch
