+++ 00_Base/src/index.ts
@@ -18,7 +18,7 @@
 
 export { AbstractModuleApi, AsDataEndpoint, AsMessageEndpoint, HttpMethod, IModuleApi } from './interfaces/api';
 export { CacheNamespace, ICache } from './interfaces/cache/cache';
-export { AbstractCentralSystem, AbstractCentralSystemApi, AsAdminEndpoint, ClientConnection, ICentralSystem, ICentralSystemApi, IClientConnection, OcppError } from './interfaces/server';
+export { AbstractCentralSystem, ClientConnection, ICentralSystem, IClientConnection, OcppError } from './interfaces/centralsystem';
 export { AbstractMessageHandler, AbstractMessageSender, EventGroup, HandlerProperties, IMessage, IMessageConfirmation, IMessageContext, IMessageHandler, IMessageRouter, IMessageSender, Message, MessageOrigin, MessageState } from './interfaces/messages';
 export { AbstractModule, AsHandler, IModule } from './interfaces/modules';
 export { Call, CallAction, CallError, CallResult, ErrorCode, MessageTypeId } from './ocpp/rpc/message';
