+++ 00_Base/src/interfaces/modules/AbstractModule.ts
@@ -25,7 +25,7 @@ import { SystemConfig } from "../../config/types";
 import { CallAction } from "../../ocpp/rpc/message";
 import { RequestBuilder } from "../../util/request";
 import { CacheNamespace, ICache } from "../cache/cache";
-import { ClientConnection } from "../server";
+import { ClientConnection } from "../centralsystem";
 import { EventGroup, HandlerProperties, IMessage, IMessageConfirmation, IMessageHandler, IMessageSender, MessageOrigin, MessageState } from "../messages";
 
 export abstract class AbstractModule implements IModule {
