--- Swarm/src/server/api.ts
@@ -1,51 +1,0 @@
-/**
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Copyright (c) 2023 S44, LLC
- */
-
-import { FastifyInstance, FastifyRequest } from "fastify";
-import { Logger, ILogObj } from "tslog";
-import { EventGroup, HttpMethod, AbstractCentralSystemApi, AsAdminEndpoint, SystemConfig } from "@citrineos/base";
-import { CentralSystemImpl } from "./server";
-/**
- * Implementation of the central system api
- */
-export class CentralSystemApiImpl extends AbstractCentralSystemApi<CentralSystemImpl> {
-
-    /**
-     * Constructs a new instance of the class.
-     *
-     * @param {CentralSystemImpl} CentralSystemImpl - The CentralSystem implementation.
-     * @param {FastifyInstance} server - The Fastify server instance.
-     * @param {Logger<ILogObj>} [logger] - The logger instance.
-     */
-    constructor(CentralSystem: CentralSystemImpl, server: FastifyInstance, logger?: Logger<ILogObj>) {
-        super(CentralSystem, server, logger);
-    }
-    /**
-     * Admin Endpoints
-     */
-
-    @AsAdminEndpoint(EventGroup.General, HttpMethod.Put)
-    async putSystemConfig(request: FastifyRequest<{ Body: SystemConfig }>): Promise<any> {
-        this._centralSystem.config = request.body;
-        return { success: true };
-    }
-
-    @AsAdminEndpoint(EventGroup.General, HttpMethod.Get)
-    async getSystemConfig(request: FastifyRequest<{}>): Promise<any> {
-        return this._centralSystem.config;
-    }
-
-}
